daemon: Use traceback.format_exc() to get the error detail
`print(f"[__run]: Error: {e}")` doesn't give enough information for
debugging because it doesn't show the traceback, files, line numbers,
etc. It only shows the error message raised by an exception. This is
severely lacking and very bad for debugging experience.

Use `traceback.format_exc()` instead to get a better log because it
shows a complete traceback of the error.

Signed-off-by: Muhammad Rizki <riskimuhammmad1@gmail.com>
Link: https://lore.gnuweeb.org/gwml/20220721232938.503-7-kiizuha@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>