telegram: Refactor the Telegram bot
Move the Telegram bot runner script outside the telegram directory and
rename it to `tg.py`. In this state, adapt the import statement for the
affected Telegram bot to avoid breakage after the main file movement.

Reason:
There will be a Discord bot with the `dc.py` runner script in the same
directory as `tg.py`. With this, both can share the same modules that
reduce code duplication.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20220925211427.412-6-kiizuha@gnuweeb.org
Cc: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Cc: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Cc: GNU/Weeb Mailing List <gwml@vger.gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>