discord: Add error handler on events
Add an error handler to ignore the `CommandNotFound` error while someone
sends a message with the bot prefix (e.g `$hai`) that is not exist in
the bot basic commands. Actually, this error wouldn't appear if set
log_handler to None in the main file on the client.run() parameter.
But I will keep this if using a log handler in the future use.

The error log says:
discord.ext.commands.bot: Ignoring exception in command None
discord.ext.commands.errors.CommandNotFound: Command "w" is not found

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20220925211427.412-11-kiizuha@gnuweeb.org
Cc: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Cc: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Cc: GNU/Weeb Mailing List <gwml@vger.gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>