Fix the storage management after the refactor was happened
I didn't test the Telegram bot when the refactor happened. I forgot to
update the utility function, like creating a file or patch directory
into the `storage` directory, which is so important.

Add the `platform` parameter to the necessary functions and add
quote_reply() for the Discord bot to create a vertical line reply
display. The `platform` parameter is to tell it for `telegram` or
`discord`. This will place the patch into its own `storage` directory.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20221003235230.1824-3-kiizuha@gnuweeb.org
Cc: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Cc: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Cc: GNU/Weeb Mailing List <gwml@vger.gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>