daemon: telegram: db: Allow the caller to reconnect
The daemon is totally unusable when the MySQL server is restarted. It's
spinning in a loop with the following error:

   mysql.connector.errors.OperationalError: 2013 (HY000): Lost connection to MySQL server during query

When it happens, the only way to fix the situation is: restart the
daemon manually, which is obviously not productive. Create a mechanism
in the class DB to allow the caller to reconnect. This way, the caller
can automatically reconnect without having the user restart the daemon.

v2 (comment from Muhammad Rizki):
  - Use conn.ping(reconnect=True) instead of creating a new MySQL
    connection object.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>