utils: fix: Fix charset issue for get_decoded_payload()
Fix the `LookupError` that is thrown by the `decode()` function. The
`unknown-8bit` charset is not found in the built-in `decode()` function
in Python, so we must manually fix it by using an if statement and
decoding using the `ascii` charset while forcing errors using
`surrogateescape`.

Sample: https://lore.kernel.org/all/Y8Am5wAxC48N12PE@quaddy.sgn/raw

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230118175356.1853-13-kiizuha@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>