discord: logger: Refactor all logging method
Same as the telegram refactors. We use a simple logging method by using
file config and replace all custom logger method to built-in logging
method. This method is more cleaner and less code wasteful.

The reason why we should name the logger like:

    `logging.getLogger("dscord")`

Because, if we use "discord" as the logger name, it will conflict the
discord.py's logger, that's why we need to disable the discord.py's
logger not to interrupt our logger.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230118175356.1853-16-kiizuha@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>